Initial Release of the NFL Logo illustration files
It is great to be able to release the original Illustrator files for the NFL Logos I redesigned as a learning exercise. You are free to use and modify any of the logos I have made anyway you'd like - fantasy logos, school teams, dorm emblems you name it - as long as its not for direct commercial use. All I ask in return is you attribute me (a link to my site for example) and share any modifications under the same terms. All of these logos are under a Creative Commons License By Attribution, Non-commercial, Share Aaike.

 Want a t-shirt with the Patriots logo I made? Download the logo and use it for one of the many custom t-shirt outlets, make a badge, stick it on a hat. One thing to mention though, I'm a fan of the NFL and that is why I did these logos, so more than anything I'd love to see how other fans use or modify what I've done. That is why ultimately why I'm releasing the original files. So if you do use these in anyway please let me know send an email to [max@ospreydawn.com](mailto:max@ospreydawn.com). Send me a photo of what you've done with it or if its digital send me a link to it. I may feature a few of my favourite renditions in a blog post in the future.